CourseType ve CourseGroupType ile ilgili kontroller ve safyaları referanslar atına taşı
