--- app/controllers/course_management/course_group_types_controller.rb
@@ -1,55 +1,0 @@
-# frozen_string_literal: true
-
-module CourseManagement
-  class CourseGroupTypesController < ApplicationController
-    include SearchableModule
-
-    before_action :set_course_group_type, only: %i[edit update destroy]
-    before_action :authorized?
-
-    def index
-      @course_group_types = pagy_by_search(CourseGroupType.order(:name))
-    end
-
-    def new
-      @course_group_type = CourseGroupType.new
-    end
-
-    def create
-      @course_group_type = CourseGroupType.new(course_group_type_params)
-      @course_group_type.save ? redirect_with('success') : render(:new)
-    end
-
-    def edit; end
-
-    def update
-      @course_group_type.update(course_group_type_params) ? redirect_with('success') : render(:edit)
-    end
-
-    def destroy
-      if @course_group_type.destroy
-        redirect_with('success')
-      else
-        redirect_to(course_group_types_path, alert: t('.warning'))
-      end
-    end
-
-    private
-
-    def redirect_with(message)
-      redirect_to(course_group_types_path, notice: t(".#{message}"))
-    end
-
-    def set_course_group_type
-      @course_group_type = CourseGroupType.find(params[:id])
-    end
-
-    def authorized?
-      authorize([:course_management, @course_group_type || CourseGroupType])
-    end
-
-    def course_group_type_params
-      params.require(:course_group_type).permit(:name)
-    end
-  end
-end
