+++ app/controllers/reference/course_group_types_controller.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Reference
+  class CourseGroupTypesController < ApplicationController
+    include ReferenceResource
+
+    private
+
+    def secure_params
+      params.require(:course_group_type).permit(:name)
+    end
+  end
+end
