+++ app/controllers/reference/course_types_controller.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Reference
+  class CourseTypesController < ApplicationController
+    include ReferenceResource
+
+    private
+
+    def secure_params
+      params.require(:course_type).permit(:name, :code, :min_credit)
+    end
+  end
+end
