+++ app/views/reference/course_group_types/edit.html.erb
