+++ app/views/reference/course_group_types/new.html.erb
