+++ app/views/reference/course_types/edit.html.erb
