+++ app/views/reference/course_types/new.html.erb
