+++ config/routes/course_management.rb
@@ -8,8 +8,6 @@ scope module: :course_management do
 
   resources :courses
   resources :course_groups
-  resources :course_group_types, except: :show
-  resources :course_types,       except: :show
   resources :course_unit_groups
   resources :curriculums do
     member do
